+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Version 2.6.3 (in Entwicklung)
+- Neu: Reiter Ausdruck ersetzt durch Dialog. 
 - Neu: Hefe- und Zusätzegaben können einzeln zurückgesetzt werden.
 - Neu: Knopf ersetzt die Filteroptionen in der Sudauswahl und Brauübersicht.
 - Neu: Ein zentrales Hilfefeld.